+++ numpy/core/src/common/npy_dlpack.h
@@ -13,4 +13,16 @@
 // capsule.
 #define NPY_DLPACK_INTERNAL_CAPSULE_NAME "numpy_dltensor"
 
+PyObject *
+array_dlpack(PyArrayObject *self, PyObject *const *args, Py_ssize_t len_args,
+             PyObject *kwnames);
+
+
+PyObject *
+array_dlpack_device(PyArrayObject *self, PyObject *NPY_UNUSED(args));
+
+
+NPY_NO_EXPORT PyObject *
+from_dlpack(PyObject *NPY_UNUSED(self), PyObject *obj);
+
 #endif
