+++ numpy/core/src/multiarray/multiarraymodule.c
@@ -70,7 +70,7 @@ NPY_NO_EXPORT int NPY_NUMUSERTYPES = 0;
 #include "get_attr_string.h"
 #include "experimental_public_dtype_api.h"  /* _get_experimental_dtype_api */
 
-#include "common/npy_dlpack.h"
+#include "npy_dlpack.h"
 
 /*
  *****************************************************************************
@@ -4231,162 +4231,6 @@ _reload_guard(PyObject *NPY_UNUSED(self)) {
     }
     initialized = 1;
     Py_RETURN_NONE;
-}
-
-/* used internally */
-static void array_dlpack_internal_capsule_deleter(PyObject *self)
-{
-    DLManagedTensor *managed = 
-        (DLManagedTensor *)PyCapsule_GetPointer(self, NPY_DLPACK_INTERNAL_CAPSULE_NAME);
-    if (managed == NULL) {
-        return;
-    }
-    managed->deleter(managed);
-}
-
-
-NPY_NO_EXPORT PyObject *
-from_dlpack(PyObject *NPY_UNUSED(self), PyObject *obj) {
-    PyObject *capsule = PyObject_CallMethod((PyObject *)obj->ob_type,
-            "__dlpack__", "O", obj);
-    if (capsule == NULL) {
-        return NULL;
-    }
-
-    DLManagedTensor *managed = 
-        (DLManagedTensor *)PyCapsule_GetPointer(capsule,
-        NPY_DLPACK_CAPSULE_NAME);
-
-    if (managed == NULL) {
-        Py_XDECREF(capsule);
-        return NULL;
-    }
-
-    const int ndim = managed->dl_tensor.ndim;
-    if (ndim > NPY_MAXDIMS) {
-        PyErr_SetString(PyExc_RuntimeError,
-                "maxdims of DLPack tensor is higher than the supported "
-                "maxdims.");
-        Py_XDECREF(capsule);
-        return NULL;
-    }
-
-    DLDeviceType device_type = managed->dl_tensor.device.device_type;
-    if (device_type != kDLCPU &&
-            device_type != kDLCUDAHost &&
-            device_type != kDLROCMHost &&
-            device_type != kDLCUDAManaged) {
-        PyErr_SetString(PyExc_RuntimeError,
-                "Unsupported device in DLTensor.");
-        Py_XDECREF(capsule);
-        return NULL;
-    }
-
-    if (managed->dl_tensor.dtype.lanes != 1) {
-        PyErr_SetString(PyExc_RuntimeError,
-                "Unsupported lanes in DLTensor dtype.");
-        Py_XDECREF(capsule);
-        return NULL;
-    }
-
-    int typenum = -1;
-    const uint8_t bits = managed->dl_tensor.dtype.bits;
-    const npy_intp itemsize = bits / 8;
-    switch(managed->dl_tensor.dtype.code) {
-    case kDLInt:
-        switch (bits)
-        {
-            case 8: typenum = NPY_INT8; break;
-            case 16: typenum = NPY_INT16; break;
-            case 32: typenum = NPY_INT32; break;
-            case 64: typenum = NPY_INT64; break;
-        }
-        break;
-    case kDLUInt:
-        switch (bits)
-        {
-            case 8: typenum = NPY_UINT8; break;
-            case 16: typenum = NPY_UINT16; break;
-            case 32: typenum = NPY_UINT32; break;
-            case 64: typenum = NPY_UINT64; break;
-        }
-        break;
-    case kDLFloat:
-        switch (bits)
-        {
-            case 16: typenum = NPY_FLOAT16; break;
-            case 32: typenum = NPY_FLOAT32; break;
-            case 64: typenum = NPY_FLOAT64; break;
-        }
-        break;
-    case kDLComplex:
-        switch (bits)
-        {
-            case 64: typenum = NPY_COMPLEX64; break;
-            case 128: typenum = NPY_COMPLEX128; break;
-        }
-        break;
-    }
-
-    if (typenum == -1) {
-        PyErr_SetString(PyExc_RuntimeError,
-                "Unsupported dtype in DLTensor.");
-        Py_XDECREF(capsule);
-        return NULL;
-    }
-
-    PyArray_Descr *descr = PyArray_DescrFromType(typenum);
-    if (descr == NULL) {
-        Py_XDECREF(capsule);
-        return NULL;
-    }
-
-    npy_intp shape[NPY_MAXDIMS];
-    npy_intp strides[NPY_MAXDIMS];
-
-    for (int i = 0; i < ndim; ++i) {
-        shape[i] = managed->dl_tensor.shape[i];
-        // DLPack has elements as stride units, NumPy has bytes.
-        if (managed->dl_tensor.strides != NULL)
-        {
-            strides[i] = managed->dl_tensor.strides[i] * itemsize;
-        }
-    }
-
-    char *data = (char *)managed->dl_tensor.data +
-            managed->dl_tensor.byte_offset;
-
-    PyObject *ret = PyArray_NewFromDescr(&PyArray_Type, descr, ndim, shape,
-            managed->dl_tensor.strides != NULL ? strides : NULL, data, 0, NULL);
-    if (ret == NULL) {
-        Py_XDECREF(capsule);
-        Py_XDECREF(descr);
-        return NULL;
-    }
-
-    PyObject *new_capsule = PyCapsule_New(managed,
-            NPY_DLPACK_INTERNAL_CAPSULE_NAME,
-            array_dlpack_internal_capsule_deleter);
-    if (new_capsule == NULL) {
-        Py_XDECREF(capsule);
-        Py_XDECREF(ret);
-        return NULL;
-    }
-
-    if (PyArray_SetBaseObject((PyArrayObject *)ret, new_capsule) < 0) {
-        Py_XDECREF(capsule);
-        Py_XDECREF(ret);
-        return NULL;
-    }
-
-    if (PyCapsule_SetName(capsule, NPY_DLPACK_USED_CAPSULE_NAME) < 0) {
-        Py_XDECREF(capsule);
-        Py_XDECREF(ret);
-        return NULL;
-    }
-    
-    Py_XDECREF(capsule);
-    return ret;
 }
 
 static struct PyMethodDef array_module_methods[] = {
